+++ src/lib.rs
@@ -6,19 +6,18 @@ pub mod builtin;
 pub mod error;
 pub mod globals;
 pub mod kernel;
-pub mod lexer;
 pub mod loader;
-pub mod node;
-pub mod parser;
+//pub mod node;
+pub mod parse;
 pub mod test;
-pub mod token;
+//pub mod token;
 pub mod util;
 pub mod value;
 pub mod vm;
 pub use crate::builtin::*;
 pub use crate::error::*;
 pub use crate::globals::*;
-pub use crate::parser::{LvarCollector, LvarId, ParseResult};
+pub use crate::parse::parser::{LvarCollector, LvarId, ParseResult, Parser};
 pub use crate::util::*;
 pub use crate::value::*;
 pub use crate::vm::*;
