+++ src/repl.rs
@@ -1,6 +1,6 @@
 use ansi_term::Colour::Red;
 use ruruby::error::*;
-use ruruby::parser::{LvarCollector, Parser};
+//use ruruby::parser::{LvarCollector, Parser};
 //#[macro_use]
 use ruruby::*;
