+++ src/vm/executor.rs
@@ -1,6 +1,6 @@
 use super::codegen::ContextKind;
 use crate::error::*;
-use crate::parser::*;
+//use crate::parser::*;
 
 #[cfg(feature = "perf")]
 use perf::*;
